Set `use_proxy` to `false` for Ansible provisioners
The default setting of `use_proxy = true` results in the Ansible
provisioner for Packer creating a local proxy adapter to connect to the
remote instance. This can result in a number of problems with more
recent OpenSSH versions due to changes that disable RSA keys by
default. The local proxy adapter is set up with an RSA key (we cannot
change this behavior), but with the proxy disabled the provisioner
will instead use the Packer SSH configuration which is set to use an
ED25519 key. If the provisioner is unable to determine an IP address
to connect to it will fall back to creating a local proxy adapter even
with `use_proxy = false` so there is no harm in disabling the proxy by
default.